mcp-joern/requirements.txt
python-dotenv>=1.0.1
fastmcp>=2.1.0
requests>=2.32.3
psutil>=7.2.2
httpx>=0.28.1